Changes to use/support xcode 27 and ios 27
