+++ examples/swift/hello_world/CrashPanel/CrashCatalog.swift
@@ -6,7 +6,7 @@
 // https://polyformproject.org/wp-content/uploads/2020/06/PolyForm-Shield-1.0.0.txt
 
 import Foundation
-@_implementationOnly import HelloWorldCrashSupport
+import HelloWorldCrashSupport
 
 enum CrashCategory: String, CaseIterable {
     case swiftRuntime = "Swift Runtime"
