+++ test/platform/swift/test_host/sources/SceneDelegate.swift
@@ -1,0 +1,40 @@
+// capture-sdk - bitdrift's client SDK
+// Copyright Bitdrift, Inc. All rights reserved.
+//
+// Use of this source code is governed by a source available license that can be found in the
+// LICENSE file or at:
+// https://polyformproject.org/wp-content/uploads/2020/06/PolyForm-Shield-1.0.0.txt
+
+import UIKit
+
+@objc(TestHostSceneDelegate)
+final class SceneDelegate: UIResponder, UIWindowSceneDelegate {
+    var window: UIWindow?
+
+    func scene(
+        _ scene: UIScene,
+        willConnectTo _: UISceneSession,
+        options _: UIScene.ConnectionOptions
+    ) {
+        guard let windowScene = scene as? UIWindowScene else { return }
+        let window = UIWindow(windowScene: windowScene)
+        window.rootViewController = HostViewController()
+        window.makeKeyAndVisible()
+        self.window = window
+    }
+}
+
+private final class HostViewController: UIViewController {
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        self.view.backgroundColor = .white
+
+        let label = UILabel()
+        self.view.addSubview(label)
+        label.text = "Test Host"
+        label.font = .boldSystemFont(ofSize: 50)
+        label.sizeToFit()
+        label.center = self.view.center
+    }
+}
